+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 
 # Celestra version history
 
+## Celestra v6.6.0 Angelo
+
+1. Documentation, pdf and code fixes.
+2. CUT design changes in the __unittest.html__.
+3. Rename the function `isEmptyValue();` to `isEmpty();`
+4. Remove these functions:
+
+- `deleteOwnProperty();`
+- `isCallable();`
+- `isChar();`
+- `isNumeric();`
+- `isSuperset();`
+- `setDifference();`
+- `setIntersection();`
+- `setSymmetricDifference();`
+- `setUnion();`
+
 ## Celestra v6.5.1
 
 1. Documentation, pdf and code fixes.